requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
cryptography==41.0.7
debugpy==1.8.1
decorator==5.1.1
Django
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
fake-useragent==1.4.0
google-auth==2.28.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyVirtualDisplay==3.0
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
selenium==4.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
